feat(runtime): 重构运行时上下文和服务初始化架构
- 添加Debug派生到ADBCommand、ADBCmdConv、ADBExecutor和CaptureMethod
- 将port类型从usize改为u16以适配网络端口范围
- 在AppCtx中引入runtime_ctx替换原有的多个独立上下文
- 实现子进程环境完整初始化流程包括OCR服务和ADB执行器
- 更新CapMethod枚举定义移除不必要的参数
- 使用tokio::sync::RwLock替代std::sync::RwLock以支持异步操作
- 重构DeviceCtx构造函数接受window_title参数
- 移除ScriptExecutor中的独立服务字段改用共享运行时上下文
- 实现运行时上下文中的脚本信息和任务状态管理
- 添加IPC消息相关的Deserialize派生以支持序列化传输

+++ src-tauri/src/infrastructure/devices/adb_info.rs
@@ -14,6 +14,6 @@ pub enum AdbConnectStatus {
 #[serde(rename_all = "camelCase")]
 pub struct AdbInfo {
     pub ip_addr: Ipv4Addr,
-    pub port: usize,
+    pub port: u16,
     pub states: AdbConnectStatus,
 }
